Remove edit-folio-by-form bulk action
Fixes #4746
