+++ app/components/bulk_actions_form_component.rb
@@ -30,7 +30,6 @@ class BulkActionsFormComponent < ApplicationComponent
         ['Set object rights', new_rights_job_path(search_of_druids)],
         ['Start text extraction', new_text_extraction_job_path(search_of_druids)],
         ['Edit license and rights statements', new_license_and_rights_statement_job_path(search_of_druids)],
-        ["Edit #{CatalogRecordId.label}s and barcodes", new_catalog_record_id_and_barcode_job_path(search_of_druids)],
         ["Refresh MODS from #{CatalogRecordId.label}", new_refresh_mods_job_path(search_of_druids)],
         ['Set content type', new_content_type_job_path(search_of_druids)],
         ['Set collection', new_collection_job_path(search_of_druids)]
