--- app/views/bulk_actions/catalog_record_id_and_barcode_jobs/_new.html.erb
@@ -1,22 +1,0 @@
-<%= form_with url: catalog_record_id_and_barcode_job_path, class: 'new_bulk_action', data: { turbo_frame: '_top' } do |f| %>
-  <%= render 'bulk_actions/errors' %>
-
-  <span class='help-block'>
-    Adds or updates <%= CatalogRecordId.label %>s and/or barcodes associated with objects.
-  </span>
-  <p>You must have a druid list and a <%= CatalogRecordId.label %> and/or a barcode list. Paste each list in its own box. Each list must be the same length and updates will be applied in order by line. Use a blank line to indicate that a <%= CatalogRecordId.label %>/barcode should be deleted. To enter multiple <%= CatalogRecordId.label %>s for one druid, enter each <%= CatalogRecordId.label %> on the same line, separated by commas.</p>
-
-  <div class="mb-3">
-    <%= f.check_box :use_catalog_record_ids_option %>
-    <%= f.label :use_catalog_record_ids_option, "Update #{CatalogRecordId.label}", class: 'ms-2' %>
-    <%= f.text_area :catalog_record_ids, class: 'form-control', rows: 10, placeholder: "Check the box and leave this area blank to remove existing #{CatalogRecordId.label}s" %>
-  </div>
-  <div class="mb-3">
-    <%= f.check_box :use_barcodes_option %>
-    <%= f.label :use_barcodes_option, 'Update barcodes', class: 'ms-2' %>
-    <%= f.text_area :barcodes, class: 'form-control', rows: 10, placeholder: 'Check the box and leave this area blank to remove existing barcodes' %>
-  </div>
-
-  <%= render 'bulk_actions/druids', f: %>
-  <%= render 'bulk_actions/common_fields', f: %>
-<% end %>
