--- app/views/bulk_actions/catalog_record_id_and_barcode_jobs/new.html.erb
@@ -1,3 +1,0 @@
-<turbo-frame id="bulk-action-form">
-  <%= render 'new' %>
-</turbo-frame>
