--- app/views/bulk_actions/catalog_record_id_and_barcode_jobs/new.turbo_stream.erb
@@ -1,1 +1,0 @@
-<%= turbo_stream.replace("bulk-action-form", partial: 'new') %>
